+++ user-exec.c
@@ -82,7 +82,7 @@ void cpu_resume_from_signal(CPUState *env1, void *puc)
    the effective address of the memory exception. 'is_write' is 1 if a
    write caused the exception and otherwise 0'. 'old_set' is the
    signal set which should be restored */
-static inline int handle_cpu_signal(unsigned long pc, unsigned long address,
+static inline int handle_cpu_signal(uintptr_t pc, unsigned long address,
                                     int is_write, sigset_t *old_set,
                                     void *puc)
 {
@@ -172,7 +172,7 @@ int cpu_signal_handler(int host_signum, void *pinfo,
 #else
     struct ucontext *uc = puc;
 #endif
-    unsigned long pc;
+    uintptr_t pc;
     int trapno;
 
 #ifndef REG_EIP
@@ -219,7 +219,7 @@ int cpu_signal_handler(int host_signum, void *pinfo,
                        void *puc)
 {
     siginfo_t *info = pinfo;
-    unsigned long pc;
+    uintptr_t pc;
 #if defined(__NetBSD__) || defined(__FreeBSD__) || defined(__DragonFly__)
     ucontext_t *uc = puc;
 #elif defined(__OpenBSD__)
@@ -332,7 +332,7 @@ int cpu_signal_handler(int host_signum, void *pinfo,
 #else
     struct ucontext *uc = puc;
 #endif
-    unsigned long pc;
+    uintptr_t pc;
     int is_write;
 
     pc = IAR_sig(uc);
@@ -393,15 +393,15 @@ int cpu_signal_handler(int host_signum, void *pinfo,
     uint32_t *regs = (uint32_t *)(info + 1);
     void *sigmask = (regs + 20);
     /* XXX: is there a standard glibc define ? */
-    unsigned long pc = regs[1];
+    uintptr_t pc = regs[1];
 #else
 #ifdef __linux__
     struct sigcontext *sc = puc;
-    unsigned long pc = sc->sigc_regs.tpc;
+    uintptr_t pc = sc->sigc_regs.tpc;
     void *sigmask = (void *)sc->sigc_mask;
 #elif defined(__OpenBSD__)
     struct sigcontext *uc = puc;
-    unsigned long pc = uc->sc_pc;
+    uintptr_t pc = uc->sc_pc;
     void *sigmask = (void *)(long)uc->sc_mask;
 #endif
 #endif
@@ -445,7 +445,7 @@ int cpu_signal_handler(int host_signum, void *pinfo,
 {
     siginfo_t *info = pinfo;
     struct ucontext *uc = puc;
-    unsigned long pc;
+    uintptr_t pc;
     int is_write;
 
 #if (__GLIBC__ < 2 || (__GLIBC__ == 2 && __GLIBC_MINOR__ <= 3))
@@ -467,7 +467,7 @@ int cpu_signal_handler(int host_signum, void *pinfo,
 {
     siginfo_t *info = pinfo;
     struct ucontext *uc = puc;
-    unsigned long pc;
+    uintptr_t pc;
     int is_write;
 
     pc = uc->uc_mcontext.gregs[16];
@@ -520,7 +520,7 @@ int cpu_signal_handler(int host_signum, void *pinfo,
 {
     siginfo_t *info = pinfo;
     struct ucontext *uc = puc;
-    unsigned long pc;
+    uintptr_t pc;
     uint16_t *pinsn;
     int is_write = 0;
 
@@ -589,7 +589,7 @@ int cpu_signal_handler(int host_signum, void *pinfo,
 {
     struct siginfo *info = pinfo;
     struct ucontext *uc = puc;
-    unsigned long pc = uc->uc_mcontext.sc_iaoq[0];
+    uintptr_t pc = uc->uc_mcontext.sc_iaoq[0];
     uint32_t insn = *(uint32_t *)pc;
     int is_write = 0;
